// src/app/shared/models/card/card-types.enum.ts
export enum CardTypes {
  card = "card",
  display = "card card_type-display",
  displaySelected = "card card_type-display selected",
  displayShown = "card card_type-display-shown ",
  displayShownSelected = "card card_type-display-shown selected",
  interactive = "card card_type-interactive",
  interactiveSelected = "card card_type-interactive selected",
}